cp01/04_Gradio_Docker/requirements.txt
gradio==3.50.2
numpy==1.26.1
Pillow==10.1.0